Refactor ColumnLayout to be more modular and use fewer watchers
